# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dhruva PGRS API
  version: 1.0.0
  description: |
    Andhra Pradesh Public Grievance Redressal System (PGRS) API

    **Core Innovation**: Verification-First Architecture
    - Citizens MUST approve before case closure
    - Immutable audit trail (tamper-proof)
    - Multi-channel intake (web, mobile, WhatsApp, SMS, voice)

    **Philosophy**: Antifragile, Modular, Test-Driven Development

    ## API Versioning Policy

    **Current Version**: v1.0 (Stable)
    **Base URL**: `https://api.dhruva.ap.gov.in/v1`

    ### Version Support Lifecycle

    | Version | Status        | Launch Date | Deprecation Date | Sunset Date  |
    |---------|---------------|-------------|------------------|--------------|
    | v1.0    | Stable        | 2025-12-01  | TBD              | TBD          |
    | v2.0    | Planned       | 2026-06-01  | -                | -            |

    ### Support Phases

    1. **Stable** (Current)
       - Fully supported with bug fixes and security patches
       - Backward-compatible changes only
       - No breaking changes

    2. **Deprecated** (When v2.0 launches)
       - v1.0 remains operational for 12 months after v2.0 launch
       - Security patches only (no new features)
       - Response headers include deprecation warnings
       - Clients encouraged to migrate to v2.0

    3. **Sunset** (12 months after deprecation)
       - v1.0 API shut down
       - All requests return HTTP 410 Gone
...

    ### Deprecation Headers

...

...

...

...
servers:
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Officer login
      description: |
        Authenticate officer and return JWT token.

        **Performance Budget**: p95 < 500ms
        **Rate Limit**: 5 attempts per minute per IP
      operationId: loginOfficer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: "officer_vijayawada"
              password: "password123"
      responses:
        '200':
          description: Login successful
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                token_type: "bearer"
                expires_in: 3600
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  username: "officer_vijayawada"
                  full_name: "Ramesh Kumar"
                  role: "officer"
                  department:
                    id: 1
                    code: "REV"
                    name: "Revenue Department"
                  district:
                    id: 5
                    code: "05"
                    name: "Krishna"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Unauthorized"
                message: "Invalid username or password"
                status_code: 401
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: Officer logout
      description: |
        Logout officer and invalidate token.

        **Performance Budget**: p95 < 200ms
      operationId: logoutOfficer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: |
        Get currently authenticated user details.

        **Performance Budget**: p95 < 200ms
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ==========================================
  # GRIEVANCE ENDPOINTS
  # ==========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - status_code
      properties:
        error:
          type: string
          description: Error type (e.g., "Validation Error", "Unauthorized")
          example: "Validation Error"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid phone number format"
        status_code:
          type: integer
          description: HTTP status code
          example: 400
        details:
          type: object
          description: Additional error details (validation errors, etc.)
          additionalProperties: true
        retry_after:
          type: integer
          description: Seconds until retry allowed (for rate limits)
          example: 60
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          example: "officer_vijayawada"
        password:
          type: string
          format: password
          minLength: 8
          example: "password123"
    LoginResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          enum: [bearer]
          example: "bearer"
        expires_in:
          type: integer
          description: Token expiry time in seconds
          example: 3600
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      required:
        - id
        - username
        - full_name
        - role
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        username:
          type: string
          example: "officer_vijayawada"
        email:
          type: string
          format: email
          example: "officer.vijayawada@dhruva.ap.gov.in"
        full_name:
          type: string
          example: "Ramesh Kumar"
        phone:
          type: string
          pattern: '^\+91[6-9]\d{9}$'
          example: "+919876543211"
        role:
          type: string
          enum: [officer, supervisor, admin]
          example: "officer"
        department:
          $ref: '#/components/schemas/DepartmentResponse'
        district:
          $ref: '#/components/schemas/DistrictResponse'
        is_active:
          type: boolean
          example: true
        last_login_at:
          type: string
          format: date-time
          nullable: true

    # ==========================================
    # GRIEVANCE SCHEMAS
    # ==========================================
...
    DistrictResponse:
      type: object
      required:
        - id
        - code
        - name
      properties:
        id:
          type: integer
          example: 5
        code:
          type: string
          pattern: '^[0-9]{2}$'
          example: "05"
        name:
          type: string
          example: "Krishna"
    DepartmentResponse:
      type: object
      required:
        - id
        - code
        - name
        - sla_days
      properties:
        id:
          type: integer
          example: 2
        code:
          type: string
          example: "HLTH"
        name:
          type: string
          example: "Health Department"
        name_telugu:
          type: string
          nullable: true
          example: "ఆరోగ్య శాఖ"
        description:
          type: string
          nullable: true
        sla_days:
          type: integer
          description: Default SLA in days
          example: 7

    # ==========================================
    # ATTACHMENT, AUDIT, VERIFICATION SCHEMAS
    # ==========================================
...
  responses:
...
    UnauthorizedError:
      description: Unauthorized (invalid or missing token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Invalid or expired token"
            status_code: 401
...
    TooManyRequests:
      description: Too many requests (rate limit exceeded)
      headers:
        X-RateLimit-Limit:
          description: Maximum requests allowed in current window
          schema:
            type: integer
            example: 100
        X-RateLimit-Remaining:
          description: Requests remaining in current window
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          description: Unix timestamp when rate limit resets
          schema:
            type: integer
            format: int64
            example: 1732450800
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
            example: 45
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Too Many Requests"
            message: "Rate limit exceeded: 100 requests per minute. Try again in 45 seconds."
            status_code: 429
...
  headers:
    X-RateLimit-Limit:
      description: Maximum requests allowed in current window
      schema:
        type: integer
      example: 100
    X-RateLimit-Remaining:
      description: Requests remaining in current window
      schema:
        type: integer
      example: 95
    X-RateLimit-Reset:
      description: Unix timestamp (seconds since epoch) when rate limit window resets
      schema:
        type: integer
        format: int64
      example: 1732450800